Improved ConvertXmlObject.cs / Bin Batch
ConvertXmlObject.cs now uses a proper calculation method to convert rotation matrices to unity quaternion.
(This is not tested with every single object, but should work nonetheless)
Notes: This only convert images at the moment, more of the trigger and areas will be coming soon.

Bin Compile and Decompiling should now have a separate option to compile or decompile the entire folder instead of single one.